--- app/src/main/java/org/sopt/app/interfaces/alert/AlertRepository.java
@@ -1,7 +1,0 @@
-package org.sopt.app.interfaces.alert;
-
-import org.sopt.app.domain.entity.Alert;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AlertRepository extends JpaRepository<Alert, Integer> {
-}
